+++ executor.go
@@ -5,5 +5,5 @@ type executor interface {
 	getPushUrls(id string, options *options) (urls []Url, err error)
 	getPullCameras(id string, options *options) (cameras []Camera, err error)
 	stop(id string, options *options) (success bool, err error)
-	getViewerNum(id string, options *options) (viewerNum int, err error)
+	getViewerNum(id string, options *options) (viewerNum int64, err error)
 }
