+++ migu.go
@@ -173,7 +173,7 @@ func (m *migu) getRecordUrls(id string, options *options) (cameras []Camera, err
 	return
 }
 
-func (m *migu) getViewerNum(id string, options *options) (viewerNum int, err error) {
+func (m *migu) getViewerNum(id string, options *options) (viewerNum int64, err error) {
 	pullReq := &miguStreamReq{
 		ChannelId: id,
 	}
@@ -183,7 +183,7 @@ func (m *migu) getViewerNum(id string, options *options) (viewerNum int, err err
 		return
 	}
 
-	viewerNum = pullRsp.Result.ViewerNum
+	viewerNum = int64(pullRsp.Result.ViewerNum)
 
 	return
 }
