+++ tencentyun.go
@@ -90,7 +90,7 @@ func (t *tencentyun) stop(_ string, _ *options) (success bool, err error) {
 	return
 }
 
-func (t *tencentyun) getViewerNum(id string, options *options) (viewerNum int, err error) {
+func (t *tencentyun) getViewerNum(id string, options *options) (viewerNum int64, err error) {
 	return 0, nil
 }
